Add .gitignore, enhance FileReader, and implement main functionality for Zmogus class
