+++ zmoniu_sarasas.txt
@@ -1,0 +1,1 @@
+ana 18 metų.
